requirements.txt
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
numpy==2.1.2
defusedxml==0.7.1
Flask==3.0.3
pandas==2.2.3
psycopg2-binary==2.9.10
bokeh==3.4.3
gunicorn==23.0.0
SQLAlchemy==2.0.36
